requirements-dev.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.6
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-backend-kitty==2.1.2
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pycairo==1.27.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
